fix: avoid integer truncation/overflow for large encode payloads
Found during a source audit; both only bite at the multi-GB boundary that the
streaming encode work newly makes reachable:

- FileEncodingPlan.fileSize is now long; the raw byte count was cast to int, so a
  >2GB (e.g. highly compressible) source produced a negative size in the manifest
  and corrupted EncodeService's totalOrigBytes accumulation.
- EncodeService chunk-end offsets use long math: (i+1)*chunkDataSize (and the
  reencode equivalent) could overflow int just below the ~2GB encodedSize cap
  before Math.min clamps it, yielding a wrong read window.

+++ apps/sender/src/main/java/airbridge/sender/EncodeService.java
@@ -137,7 +137,9 @@ final class EncodeService {
                     for (int i = 0; i < plan.totalChunks(); i++) {
                         throwIfCancelled(effectiveCancelled);
                         int start = i * chunkDataSize;
-                        int end = Math.min((i + 1) * chunkDataSize, plan.encodedSize());
+                        // long math: (i + 1) * chunkDataSize can exceed Integer.MAX_VALUE for
+                        // payloads near the 2GB encodedSize cap before Math.min clamps it.
+                        int end = (int) Math.min((long) (i + 1) * chunkDataSize, plan.encodedSize());
                         String chunkData = plan.readChunk(start, end);
 
                         String payload = QrPayloadSupport.buildPayload(
@@ -263,7 +265,7 @@ final class EncodeService {
                     }
 
                     int start = (chunkIdx - 1) * chunkDataSize;
-                    int end = Math.min(chunkIdx * chunkDataSize, plan.encodedSize());
+                    int end = (int) Math.min((long) chunkIdx * chunkDataSize, plan.encodedSize());
                     String chunkData = plan.readChunk(start, end);
 
                     String payload = QrPayloadSupport.buildPayload(
